make luma bias work

+++ av1/encoder/aq_variance.h
@@ -22,7 +22,7 @@ extern "C" {
 #if !CONFIG_REALTIME_ONLY
 void av1_vaq_frame_setup(AV1_COMP *cpi);
 
-int av1_log_block_avg(MACROBLOCK *x, BLOCK_SIZE bs);
+float av1_log_block_avg(MACROBLOCK *x, BLOCK_SIZE bs);
 int av1_compute_q_from_energy_level_deltaq_mode(const AV1_COMP *const cpi,
                                                 int block_var_level);
 int av1_block_wavelet_energy_level(const AV1_COMP *cpi, MACROBLOCK *x,
